Addition of ViteJS and Deletion of Unused Directories
The project has been updated to include the ViteJS framework, in order to improve its performance and overall development experience. As part of this process, certain directories that were no longer needed or used were deleted from the codebase. This refactor will result in a more streamlined and efficient codebase, allowing for faster and more agile development moving forward.